requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
boto3==1.35.99
python-dotenv==1.2.2